requirements.txt
Flask>=3.1.1
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
python-dateutil>=2.9.0.post0
requests>=2.32.0
openai==1.3.0
Werkzeug>=3.0.0
SQLAlchemy==2.0.21
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.2.0
blinker==1.9.0
typing_extensions==4.8.0
six==1.16.0
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
urllib3==2.0.6
distro==1.8.0
httpcore==1.0.9
httpx==0.27.2
anyio==3.7.1
sniffio==1.3.0
h11>=0.15.0
tqdm==4.66.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
async-timeout==4.0.3
attrs==23.1.0
frozenlist==1.4.0
multidict==6.0.4
propcache==0.3.2
yarl==1.20.1
gunicorn>=21.2

fpdf2>=2.7.0
